+++ pkg/reconciler/v1alpha1/broker/resources/subscription.go
@@ -1,0 +1,42 @@
+package resources
+
+import (
+	"fmt"
+
+	"github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+)
+
+// NewSubscription returns a placeholder subscription for trigger 't', channel 'c', and service 'svc'.
+func NewSubscription(b *v1alpha1.Broker, c *v1alpha1.Channel, svc *corev1.Service) *v1alpha1.Subscription {
+	return &v1alpha1.Subscription{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace:    b.Namespace,
+			GenerateName: fmt.Sprintf("internal-ingress-%s-", b.Name),
+			OwnerReferences: []metav1.OwnerReference{
+				*metav1.NewControllerRef(b, schema.GroupVersionKind{
+					Group:   v1alpha1.SchemeGroupVersion.Group,
+					Version: v1alpha1.SchemeGroupVersion.Version,
+					Kind:    "Broker",
+				}),
+			},
+			Labels: IngressSubscriptionLabels(b),
+		},
+		Spec: v1alpha1.SubscriptionSpec{
+			Channel: corev1.ObjectReference{
+				APIVersion: v1alpha1.SchemeGroupVersion.String(),
+				Kind:       "Channel",
+				Name:       c.Name,
+			},
+			Subscriber: &v1alpha1.SubscriberSpec{
+				Ref: &corev1.ObjectReference{
+					APIVersion: "v1",
+					Kind:       "Service",
+					Name:       svc.Name,
+				},
+			},
+		},
+	}
+}
